+++ CHANGES.md
@@ -34,8 +34,8 @@
   - この修正は Sora との内部的なやり取り部分にのみ影響するため、SDK ユーザーへの影響はない
   - @zztkm
 - [UPDATE] `SoraMediaChannel.Listener` に Sora から切断されたときのステータスコードと理由を取得できる `onClose` を追加する
-  - Sora からの切断結果を表す `SoraCloseResult` を追加した
-  - WebSocket シグナリングの切断結果を表す `SignalingChannelDisconnectResult` を追加した
+  - Sora から切断した時に通知されるイベントである `SoraCloseEvent` を追加した
+  - WebSocket シグナリング切断時に通知されるイベントである `SignalingChannelCloseEvent` を追加した
   - @zztkm
 - [UPDATE] `SignalingChannelImpl` の `WebSocketListener.onClosed` の処理で WebSocket ステータスコードが 1000 以外の場合に onError を呼び出さないようにする
   - 2025.1.0 まで onError のコールバック呼び出しが定義されていたが、実際は onClosing が実行された時点で SignalingChannelImpl の listener の参照が削除され、onError は呼び出されなかったため動作に変更はない